fix: types

+++ graphics/events/IMouseEvent.hpp
@@ -23,5 +23,5 @@ public:
      *
      * @return Position of the mouse
      */
-    virtual const shared::types::Vector2f getPosition(void) const noexcept = 0;
+    virtual const shared::types::Vector2i getPosition(void) const noexcept = 0;
 };
